added css file

+++ flask_app.py
@@ -19,7 +19,7 @@ def predict():
     '''
     For rendering results on HTML GUI
     '''
-    int_features = [int(x) for x in request.form.values()]
+    int_features = [float(x) for x in request.form.values()]
     final_features = (np.array(int_features)).reshape(1,-1)
     prediction = model.predict(final_features)
     if(prediction==0):
